requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
anywidget==0.9.18
appdirs==1.4.4
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.6
choreographer==1.2.0
click==8.2.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
docutils==0.22
et_xmlfile==2.0.0
executing==2.2.0
faicons==0.2.2
fonttools==4.59.1
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaleido==1.1.0
kiwisolver==1.4.9
linkify-it-py==2.0.3
logistro==2.0.1
loro==1.5.4
marimo==0.15.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.2.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
openai==1.109.1
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
psygnal==0.14.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
PyGithub==2.9.0
Pygments==2.19.2
PyJWT==2.12.1
pymdown-extensions==10.16.1
PyNaCl==1.6.2
pyparsing==3.2.3
pyprojroot==0.3.0
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
questionary==2.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.16.1
setuptools==80.9.0
shiny==1.4.0
shinywidgets==0.7.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.3
statsmodels==0.14.5
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
UpSetPlot==0.9.0
urllib3==2.6.3
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
